Исправление текста, введение обсужденных стандартов кода
