--- src/js/common/layout/components/footer-links-group/index.js
@@ -1,1 +1,0 @@
-export { default as FooterLinksGroup } from './footer-links-group';
