--- src/js/common/layout/components/footer/footer.constants.js
@@ -1,54 +1,0 @@
-import t from '../../../../constants/texts';
-
-export const USEFUL_LINKS = [
-  {
-    label: t.footer.repositoryLabel,
-    link: 'https://github.com/TheDrHax/BlackSilverUfa/',
-    iconClass: 'fa-github',
-  },
-  {
-    label: t.footer.channelLabel,
-    link: 'https://t.me/BlackUFA_Monitor',
-    iconClass: 'fa-telegram',
-  },
-  {
-    label: t.footer.denialLabel,
-    link: 'https://matomo.thedrhax.pw/index.php?module=CoreAdminHome&action=optOut&language=ru',
-  },
-];
-
-export const CHANNEL_LINKS = [
-  {
-    label: 'BlackSilverUFA',
-    link: 'https://www.youtube.com/user/BlackSilverUFA',
-    iconClass: 'fa-youtube',
-  },
-  {
-    label: 'BlackSilverChannel',
-    link: 'https://www.youtube.com/user/BlackSilverChannel',
-    iconClass: 'fa-youtube',
-  },
-  {
-    label: 'BlackUFA',
-    link: 'https://www.twitch.tv/blackufa/',
-    iconClass: 'fa-twitch',
-  },
-];
-
-export const NETWORK_LINKS = [
-  {
-    label: t.footer.vkgroupLabel,
-    link: 'https://vk.com/b_silver',
-    iconClass: 'fa-vk',
-  },
-  {
-    label: t.footer.vkLabel,
-    link: 'https://vk.com/blacksilverufa',
-    iconClass: 'fa-vk',
-  },
-  {
-    label: 'Twitter',
-    link: 'https://twitter.com/Sempai_Black',
-    iconClass: 'fa-twitter',
-  },
-];
