--- src/js/common/layout/components/footer/index.js
@@ -1,1 +1,0 @@
-export { default as Footer } from './footer';
