--- src/js/common/layout/components/header-quick-search/header-quick-search.constants.js
@@ -1,2 +1,0 @@
-export const MIN_CHARS = 2;
-export const MAX_CHARS = 20;
