--- src/js/common/layout/components/header-quick-search/index.js
@@ -1,1 +1,0 @@
-export { default as HeaderQuickSearch } from './header-quick-search';
