--- src/js/common/layout/components/header/index.js
@@ -1,1 +1,0 @@
-export { default as Header } from './header';
