--- src/js/common/layout/components/index.js
@@ -1,4 +1,0 @@
-export * from './footer';
-export * from './footer-links-group';
-export * from './header';
-export * from './header-quick-search';
